+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- quantityPerPage and quantityFirstPage on Reviews
+- `default ordination type` to the app's settings
 
 ## [1.2.2] - 2019-09-05
 ### Fixed